Refactor for improved type safety
This is a substantial refactoring with the primary goal to reduce
reliance on duck typing by using more concrete types in order to
benefit from static type checking.

The Reference object has been broken up into untyped and typed
references, where typed refs are subclasses of Reference, and which
implements the various data needed for rendering.

This also begins paving the way to supporting multiple renderers, which
will be needed for #5.

In order to use slightly more recent type hinting features, Python 3.8
is now required.  That makes this commit a breaking change, meaning the
next release will require a major version bump.